[SR][Code cleanup] Typedef/default for kwargs
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69164

We have lots of methods that take `std::unordered_map<std::string, c10::IValue>` now. That's kind of ugly and cumbersome to type, so add a `KWargs` typedef.

Also made the `operator()` default `kwargs` to empty. Note that we could have another overload that doesn't take `kwargs` at all, but the perf gain is so minuscule it's probably not worth it.
ghstack-source-id: 144581608

Differential Revision: [D32734677](https://our.internmc.facebook.com/intern/diff/D32734677/)